Splitting Gooey code out for slim cmdline build, Dockerfile for linux builds
