+++ docker/Dockerfile.build
@@ -1,0 +1,21 @@
+# syntax=docker/dockerfile:1
+
+FROM python:3.9-bullseye
+
+WORKDIR /app
+
+RUN apt update
+RUN apt install -y libpq-dev build-essential
+
+COPY requirements_no_gui.txt requirements_no_gui.txt
+
+RUN pip3 install pip --upgrade
+RUN pip3 install -r requirements_no_gui.txt
+
+COPY stats/ /app/stats
+
+COPY scripts/release/build_no_gui.sh .
+
+RUN chmod +x build_no_gui.sh
+
+CMD ["./build_no_gui.sh"]
