+++ requirements_no_gui.txt
@@ -1,0 +1,15 @@
+requests==2.26.0
+certifi==2020.12.5
+cffi==1.14.5
+chardet==4.0.0
+cryptography
+idna==2.10
+peewee==3.14.8
+pycparser==2.20
+PyMySQL==1.0.0
+python-dotenv==0.15.0
+six==1.15.0
+urllib3==1.26.5
+psycopg2==2.9.1
+psycopg2-binary==2.9.1
+pyinstaller
