+++ scripts/release/build_no_gui.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+pyinstaller --paths=stats stats/nba_sql.py -F
