+++ stats/args.py
@@ -1,4 +1,3 @@
-from gooey import GooeyParser
 from utils import generate_valid_seasons
 
 """
@@ -6,26 +5,10 @@ Creates a parser.
 """
 
 
-def create_parser():
+def create_parser(parser):
     """
     Creates and returns a Gooey parser.
     """
-
-    parser = GooeyParser(description="nba-sql")
-
-    mode_parser = parser.add_mutually_exclusive_group(
-        required=True,
-        gooey_options={
-            'initial_selection': 0
-        })
-    mode_parser.add_argument(
-        '--default_mode',
-        help='Mode to create the database and load historic data. Use this mode when creating a new database or when trying to load a specific season or a range of seasons.',
-        action='store_true')
-    mode_parser.add_argument(
-        '--current_season_mode',
-        help='Mode to refresh the current season. Use this mode on an existing database to update it with the latest data.',
-        action='store_true')
 
     parser.add_argument(
         '--database',
@@ -49,23 +32,6 @@ def create_parser():
         help="Database Username (Not Needed For SQLite)",
         default=None)
 
-    parser.add_argument(
-        '--password',
-        help="Database Password (Not Needed For SQLite)",
-        widget='PasswordField',
-        default=None)
-
-    valid_seasons = generate_valid_seasons()
-    last_loadable_season = valid_seasons[-1]
-
-    parser.add_argument(
-        '--seasons',
-        dest='seasons',
-        default=[last_loadable_season],
-        choices=valid_seasons,
-        widget='Listbox',
-        nargs="*",
-        help='The seasons flag loads the database with the specified season.  The format of the season should be in the form "YYYY-YY".  The default behavior is loading the current season.')
 
     parser.add_argument(
         '--create-schema',
@@ -79,15 +45,6 @@ def create_parser():
         dest='request_gap',
         default='.7',
         help='This flag exists to prevent rate limiting, and injects the desired amount of time inbetween requesting resources.')
-
-    parser.add_argument(
-        '--skip-tables',
-        action='store',
-        nargs="*",
-        default='',
-        choices=['player_season', 'player_game_log', 'play_by_play', 'pgtt', 'shot_chart_detail', 'game', 'event_message_type', 'team', 'player', ''],
-        widget='Listbox',
-        help='Use this option to skip loading certain tables.')
 
     #To fix issue https://github.com/mpope9/nba-sql/issues/56
     parser.add_argument(
